www: Push the Jobs with multiple conditions
Natively, if you push items to the list, it will have problems, so
we have to compare the item in the state list vs Job list on cases
of cloning (by converting list to string).
And always assign, first element from the CTC List as the "task"
for the job, and for mandatory/All-task job, push as "av2-a2-2k" set.
This default assignment is used for frontend rendering.